fix done

+++ src/java/dal/CommunityPostDao.java
@@ -22,11 +22,9 @@ import java.sql.SQLException;
  */
 public class CommunityPostDao extends DBContext {
 
-
     private UserDao userDao = new UserDao();
 
-   
-
+    // hàm cho phép chúng ta tạo ra bài đăng
     public void addPost(CommunityPost cp) throws SQLException {
         String query = "INSERT INTO [dbo].[Community_post]\n"
                 + "           ([tittle]\n"
@@ -47,10 +45,12 @@ public class CommunityPostDao extends DBContext {
         }
     }
 
-    public void deletePost(int postId) {
-        String query = "DELETE FROM Community_post WHERE id = ?";
+    // Hàm này sẽ xóa tất cả những bài đăng thuộc về 1 user nào đó
+    public void deletePost(int postId, int user_id) {
+        String query = "DELETE FROM Community_post WHERE [id] = ? and [user_id] = ? ";
         try (PreparedStatement statement = connection.prepareStatement(query)) {
             statement.setInt(1, postId);
+            statement.setInt(2, user_id);
             statement.executeUpdate();
         } catch (SQLException e) {
             e.printStackTrace();
@@ -65,7 +65,7 @@ public class CommunityPostDao extends DBContext {
                 CommunityPost cp = new CommunityPost();
                 cp.setId(rs.getInt("id"));
                 cp.setContext(rs.getString("context"));
-                cp.setTitle(rs.getString("title"));
+                cp.setTitle(rs.getString("tittle"));
                 User userId = userDao.getUser(rs.getInt("user_id"));
                 cp.setUser_id(userId);
                 cp.setTime(rs.getDate("time"));
@@ -78,6 +78,68 @@ public class CommunityPostDao extends DBContext {
         return posts;
     }
 
+    // đây là list tìm kiếm có thể phân trang
+    public List<CommunityPost> searchCommunityPostsList(String title, int pageNumber, int pageSize) {
+        List<CommunityPost> list = new ArrayList<>();
+        String sql = "SELECT * FROM Community_post where 1=1  ";
+
+        if (title != null) {
+            sql += " and [tittle] like ? ";
+        }
+
+        int offset = (pageNumber - 1) * pageSize;
+
+        sql += "OFFSET " + offset + " ROWS ";
+        sql += "FETCH NEXT " + pageSize + " ROWS ONLY ";
+
+        try {
+            PreparedStatement st = connection.prepareStatement(sql);
+            st.setString(1, title);
+            ResultSet rs = st.executeQuery();
+            while (rs.next()) {
+                CommunityPost cp = new CommunityPost();
+                cp.setId(rs.getInt("id"));
+                cp.setContext(rs.getString("context"));
+                cp.setTitle(rs.getString("tittle"));
+                User userId = userDao.getUser(rs.getInt("user_id"));
+                cp.setUser_id(userId);
+                cp.setTime(rs.getDate("time"));
+                cp.setFirst_image(rs.getString("first_image"));
+                list.add(cp);
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return list;
+    }
+
+    // hàm này sẽ lấy ra được kích thước của list vừa mới search
+    public int getSizeOfListSearch(String title) {
+        List<CommunityPost> list = new ArrayList<>();
+        String sql = "SELECT COUNT([id]) AS list_size\n"
+                + "FROM [ams].[dbo].[Community_post]\n"
+                + "where 1=1 ";
+
+        if (title != null) {
+            sql += " and [tittle] like ? ";
+        }
+
+        int result = 0;
+        try {
+            PreparedStatement st = connection.prepareStatement(sql);
+            st.setString(1, title);
+            ResultSet rs = st.executeQuery();
+            if (rs.next()) {
+                result = rs.getInt("list_size");
+                return result;
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return 0;
+    }
+
+    // hàm này cho phép chúng ta có thể thay đổi thông tin của bài đăng của chính mình
     public void updatePost(CommunityPost post) {
         String query = "UPDATE Community_post SET tittle = ?, context = ? WHERE id = ?";
         try (PreparedStatement statement = connection.prepareStatement(query)) {
@@ -91,6 +153,7 @@ public class CommunityPostDao extends DBContext {
         }
     }
 
+    // lấy ra bài post mới nhất thuộc về 1 user nào đó
     public CommunityPost getNewesPost(int user_id) {
         String sql = " SELECT TOP (1)*\n"
                 + "  FROM [dbo].[Community_post] where[id] = ? ";
@@ -114,6 +177,7 @@ public class CommunityPostDao extends DBContext {
         return null;
     }
 
+    // lấy ra 1 bài đăng bất kì
     public CommunityPost getCommunityPost(int id) {
         String sql = " SELECT *\n"
                 + "  FROM [dbo].[Community_post] where[id] = ? ";
@@ -137,8 +201,7 @@ public class CommunityPostDao extends DBContext {
         return null;
     }
 
-    
-
+    // thêm ảnh vào trong bài đăng
     public void addPostImage(Post_image post_image) {
         String query = "INSERT INTO [dbo].[Image_post] ([image], [post_id]) VALUES (?, ?)";
         try (PreparedStatement statement = connection.prepareStatement(query)) {
@@ -150,6 +213,7 @@ public class CommunityPostDao extends DBContext {
         }
     }
 
+    //lấy ra bức ảnh đầu tiên thuộc về 1 bài đăng nào đó
     public Post_image getFirstPostImage(int post_id) {
         String query = "SELECT top 1 *\n"
                 + "  FROM [dbo].[Image_post] where [post_id] = ? order by id desc";
